Add prototypesupport of Arrow Dataset
function collect_to_arrow downloads the dataset as an Arrow table.
collect_to_dataframe convert the arrow table to julia DataFrame and
collect_to_tuples converts it to a list of named tuples.

create_df takes a julia Table, transfers is to Spark and creates new Dataset

Currently, the data is transferred by sending byte array through jcall, probably
not a very efficient way to it.